WIP: Make interfaces ADT-value pairs instead of typeclass-type.
In this commit we are using known and, as per opinions of some professionals,
correct approach to defining an interface in a Haskell-like language. (I
personally believe that typeclasses aren't interfaces and people should avoid
writing typeclasses for things that aren't fundamental). Besides, we partially
fix orphan instance issue that is currently present in purescript-simple-dom
and prevent building it with 0.7.3.

We also bump dependencies.

This commit is WIP, meaning that

 1. The project doesn't build yet and it involves a lot of tinkering to make it
    work, so before proceeding we would like to collect input from the package
    maintainer and/or community considering this approach.

 2. Current bower.json fetches (some) dependencies from strange places (such as
    our fork of purescript-dom)

If this approach will be considered feasible, we intend to finish work on
refactoring purescript-simple-dom very quickly.